+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/implementation/directconnectivity/ReflectionUtils.java
@@ -49,6 +49,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.time.Duration;
 import java.util.List;
+import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -390,5 +391,15 @@ public class ReflectionUtils {
 
     public static void setClientTelemetryMetadataHttpClient(ClientTelemetry clientTelemetry, HttpClient HttpClient) {
         set(clientTelemetry, HttpClient, "metadataHttpClient");
+    }
+
+    @SuppressWarnings("unchecked")
+    public static AtomicReference<Uri.HealthStatus> getHealthStatus(Uri uri) {
+        return get(AtomicReference.class, uri, "healthStatus");
+    }
+
+    @SuppressWarnings("unchecked")
+    public static Set<Uri.HealthStatus> getReplicaValidationScopes(GatewayAddressCache gatewayAddressCache) {
+        return get(Set.class, gatewayAddressCache, "replicaValidationScopes");
     }
 }