+++ sdk/core/azure-core-amqp/CHANGELOG.md
@@ -4,15 +4,25 @@
 
 ### Features Added
 
-- Added AMQP-level metrics to reporting number of sent batches, duration of network call, number of received message and consumer lag as well as 
-  error counters. Metrics are off by default and can be enabled with [azure-core-metrics-opentelemetry](https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/core/azure-core-metrics-opentelemetry/README.md)
-  plugin. ([#30583](https://github.com/Azure/azure-sdk-for-java/pull/30583))
-
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 2.7.1 (2022-09-01)
+
+### Features Added
+
+- Added AMQP-level metrics to reporting number of sent batches, duration of network call, number of received message and consumer lag as well as 
+  error counters. Metrics are off by default and can be enabled with [azure-core-metrics-opentelemetry](https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/core/azure-core-metrics-opentelemetry/README.md)
+  plugin. ([#30583](https://github.com/Azure/azure-sdk-for-java/pull/30583))
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded Reactor from `3.4.21` to `3.4.22`.
 
 ## 2.7.0 (2022-08-05)